Add enemy locations to biome search
