+++ src/screens/LevelsScreen.js
@@ -1,5 +1,4 @@
 import React from "react";
-import _get from "lodash.get";
 
 // Screens.
 import EnemiesScreen from "./Levels/EnemiesScreen";
